Move jailbreak guardrails into single pipeline class
We've got 3 classes for Jailbreak guardrails. It's a fairly simple
class and can just be moved into a single class in the answer composition
pipeline.

I did consider leaving the JailbreakChecker class as a separate class, but
it just didn't seem like there was a lot of value.

One thing i noticed while merging this, is that we create a ResponseError
class in the JailbreakChecker, but we don't actually raise it anywhere
so it was redundant. I've removed it, but if we want we can check
that the llm response either returns the pass or fail value and
create the object in the db with the correct status if it doesn't.

We'd need to add some additional prompt config for the fail value though
in order to do that so i've avoided it for now.

We will need to do some follow up work to update the evaluation repo
since we just return the serialised answer now so it'll need to grab
the info it wants from the serialised answer instead of the result that
was previously returned from the JailbreakChecker.
