--- spec/lib/guardrails/jailbreak_checker_spec.rb
@@ -1,76 +1,0 @@
-RSpec.describe Guardrails::JailbreakChecker, :aws_credentials_stubbed do
-  let(:input) { "User question" }
-  let(:pass_value) { "PassValue" }
-
-  before do
-    allow(described_class).to receive_messages(pass_value:)
-  end
-
-  it "calls the Claude jailbreak checker by default" do
-    result = {
-      llm_guardrail_result: pass_value,
-      llm_response: {
-        "message" => { "content" => pass_value },
-        "finish_reason" => "stop",
-        "index" => 0,
-      },
-      llm_token_usage: {
-        "prompt_tokens" => 100,
-        "completion_tokens" => 2,
-        "total_tokens" => 102,
-      },
-    }
-
-    allow(Guardrails::Claude::JailbreakChecker).to receive(:call).and_return(result)
-    described_class.call(input)
-    expect(Guardrails::Claude::JailbreakChecker).to have_received(:call).with(input)
-  end
-
-  it "returns a result object" do
-    stub_claude_jailbreak_guardrails(input)
-
-    result = described_class.call(input)
-
-    expect(result)
-      .to be_an_instance_of(described_class::Result)
-      .and have_attributes(
-        triggered: boolean,
-        llm_response: hash_including(:content, :stop_reason),
-        llm_prompt_tokens: be_a(Integer),
-        llm_completion_tokens: be_a(Integer),
-        llm_cached_tokens: be_a(Integer).or(be_nil),
-        model: be_a(String),
-      )
-  end
-
-  it "returns a result object with triggered true when guardrails fail" do
-    allow(Guardrails::Claude::JailbreakChecker).to receive(:call).with(input).and_call_original
-    stub_claude_jailbreak_guardrails(input, triggered: true)
-    second_input = "Second user question"
-    allow(Guardrails::Claude::JailbreakChecker)
-      .to receive(:call)
-      .with(second_input)
-      .and_return({
-        llm_response: {
-          "message" => { "content" => "unexpected" },
-          "finish_reason" => "stop",
-          "index" => 0,
-        },
-        llm_guardrail_result: "unexpected",
-        llm_prompt_tokens: 10,
-        llm_completion_tokens: 10,
-        llm_cached_tokens: nil,
-      })
-
-    first_result = described_class.call(input)
-    second_result = described_class.call(second_input)
-
-    expect(first_result).to have_attributes(triggered: true)
-    expect(second_result).to have_attributes(triggered: true)
-  end
-
-  it "returns a result object with triggered false when guardrails pass" do
-    stub_claude_jailbreak_guardrails(input, triggered: false)
-    expect(described_class.call(input)).to have_attributes(triggered: false)
-  end
-end
